[DRAFT]: adding attention prior, works for fp32
Adding attention prior. Cross attention scores are recomputed in the model,
stored in the runtime buffer. At each decoder step, a focus index is computed
for each request and stored in the llm request. Then it is used in transformer
buffers to create an attention mask.

The commit currently works for fp32, but not fp16.

Possible optimizations:
1) do not recompute q_proj and kv_proj,
reuse them from within attention in model definition.
2) do not copy entire scores to CPU, copy only region to search focus index in
3) introduce a kernel that creates an attention mask on GPU based on scores,
to avoid any copying at all

Signed-off-by: Viacheslav Klimkov <vklimkov@nvidia.com>